feat: Added project_config_manager interface (#572)
* Added project_config_manager interface

* Add comments

* Rebase and fix spacing

* Improcporate comments and clean up imports
